projection_diff() now works properly with uncertainties and a default radius
An integer overflow was previous causing autodetection of focus_radius to fail.
This patch uses a float, which resolves the problem. This was happening:

  $ mrcal-convert-lensmodel --viz LENSMODEL_OPENCV5 camera0.cameramodel

  RMS error of the solution: 0.14429250923230258 pixels.
  Transformation cam_fitted <-- cam_input:  rotation: 0.000 degrees, translation: [ 4.04e-06  2.90e-06 -9.86e-06] m
  Wrote 'camera0-LENSMODEL_OPENCV5.cameramodel'
  mrcal/mrcal/model_analysis.py:215: RuntimeWarning: overflow encountered in scalar multiply
    i = nps.norm2(q_off_center) < focus_radius*focus_radius
  Traceback (most recent call last):
    File "mrcal/mrcal-convert-lensmodel", line 906, in <module>
      plot,_ = mrcal.show_projection_diff( (m, m_to),
               ~~~~~~~~~~~~~~~~~~~~~~~~~~^^^^^^^^^^^^
                                           implied_Rt10      = implied_Rt10,
                                           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
      ...<6 lines>...
                                           terminal          = args.terminal,
                                           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
                                           **plotkwargs_extra)
                                           ^^^^^^^^^^^^^^^^^^^
    File "mrcal/mrcal/visualization.py", line 1235, in show_projection_diff
      difflen,diff,q0,Rt10 = mrcal.projection_diff(models,
                             ~~~~~~~~~~~~~~~~~~~~~^^^^^^^^
                                                   implied_Rt10      = implied_Rt10,
                                                   ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
      ...<5 lines>...
                                                   focus_center      = focus_center,
                                                   ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
                                                   focus_radius      = focus_radius)
                                                   ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
    File "mrcal/mrcal/model_analysis.py", line 1858, in projection_diff
      implied_Rt10__from_unprojections(q0,
      ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~^^^^
                                       # shape (len(distance),Nheight,Nwidth,3)
                                       ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
      ...<4 lines>...
                                       focus_center = focus_center,
                                       ^^^^^^^^^^^^^^^^^^^^^^^^^^^^
                                       focus_radius = focus_radius)
                                       ^^^^^^^^^^^^^^^^^^^^^^^^^^^^
    File "mrcal/mrcal/model_analysis.py", line 217, in implied_Rt10__from_unprojections
      raise Exception("Focus region contained too few points")
  Exception: Focus region contained too few points

+++ mrcal/model_analysis.py
@@ -1810,9 +1810,9 @@ A tuple
 
     if focus_radius < 0:
         if uncertainties is not None:
-            focus_radius = max(W,H) * 100 # whole imager
+            focus_radius = float(max(W,H) * 100.) # whole imager
         else:
-            focus_radius = min(W,H)/6.
+            focus_radius = float(min(W,H)/6.)
 
     if len(models) == 2:
         # Two models. Take the difference and call it good
